+++ CHANGELOG.md
@@ -5,13 +5,13 @@
 - Fix column layout issue - had delete code that was needed
 - Prevent excess renders when showing the details popup
 - Remove the search context and push into the main state
+- Remove the color scale context
 
 ## 0.0.16 -- 2022-09-13 23:43:52
 
 - Rework the hover and hotkey dynamics to go through the common store
   - This makes it much easier to interact with the "right" object
   - Still need to wire up a couple more
-- TODO: need to sort out the rebalance code again...
 - TODO: need to sort out mouse events on repaint
 
 ## 0.0.15 -- 2022-09-12 23:13:43